01-osint-toolkit/src/modules/social_media_osint.py
import requests
from bs4 import BeautifulSoup

class SocialMediaOSINT:
    def search_username(self, username):
        # Example Twitter scrape (concept only; real scraping must comply with TOS)
        url = f"https://twitter.com/{username}"
        try:
            response = requests.get(url)
            if response.status_code == 200:
                soup = BeautifulSoup(response.text, "html.parser")
                # You can extract user bio, followers, etc. here if visible
                return {"username": username, "profile_found": True}
        except Exception:
            pass
        return {"username": username, "profile_found": False}
