01-osint-toolkit/requirements.txt
beautifulsoup4==4.12.0
requests==2.31.0
pandas==2.0.0
dnspython==2.4.2
python-dotenv==1.0.0
whois==0.9.7
validators==0.22.0
